+++ services/web/server/src/simcore_service_webserver/computation_subscribe.py
@@ -5,41 +5,22 @@ from functools import wraps
 from pprint import pformat
 from typing import Callable, Coroutine, Dict, Optional
 
-from aiohttp import web
-from tenacity import before_sleep_log, retry, stop_after_attempt, wait_fixed
-
 import aio_pika
+from aiohttp import web
+from tenacity import retry
+
 from servicelib.application_keys import APP_CONFIG_KEY
+from servicelib.rabbitmq_utils import RabbitMQRetryPolicyUponInitialization
 from simcore_sdk.config.rabbit import eval_broker
 
-from .computation_config import (
-    APP_CLIENT_RABBIT_DECORATED_HANDLERS_KEY,
-    CONFIG_SECTION_NAME,
-)
+from .computation_config import (APP_CLIENT_RABBIT_DECORATED_HANDLERS_KEY,
+                                 CONFIG_SECTION_NAME)
 from .projects import projects_api
-from .projects.projects_exceptions import NodeNotFoundError, ProjectNotFoundError
+from .projects.projects_exceptions import (NodeNotFoundError,
+                                           ProjectNotFoundError)
 from .socketio.events import post_messages
 
 log = logging.getLogger(__file__)
-
-
-class RabbitMQRetryPolicyUponInitialization:
-    """ Retry policy upon service initialization
-    """
-
-    WAIT_SECS = 2
-    ATTEMPTS_COUNT = 20
-
-    def __init__(self, logger: Optional[logging.Logger] = None):
-        logger = logger or log
-
-        self.kwargs = dict(
-            wait=wait_fixed(self.WAIT_SECS),
-            stop=stop_after_attempt(self.ATTEMPTS_COUNT),
-            before_sleep=before_sleep_log(logger, logging.INFO),
-            reraise=True,
-        )
-
 
 def rabbit_adapter(app: web.Application) -> Callable:
     """this decorator allows passing additional paramters to python-socketio compatible handlers.
